wip on schema initial structure
