+++ redisvl/schema/__init__.py
@@ -1,0 +1,34 @@
+from redisvl.schema.fields import (
+    BaseField,
+    TagFieldSchema,
+    TextFieldSchema,
+    GeoFieldSchema,
+    NumericFieldSchema,
+    BaseVectorField,
+    HNSWVectorFieldSchema,
+    FlatVectorFieldSchema
+)
+from redisvl.schema.schema import (
+    StorageType,
+    IndexModel,
+    FieldsModel,
+    Schema,
+    SchemaGenerator
+)
+
+
+__all__ = [
+    "BaseField",
+    "TagFieldSchema",
+    "TextFieldSchema",
+    "GeoFieldSchema",
+    "NumericFieldSchema",
+    "BaseVectorField",
+    "HNSWVectorFieldSchema",
+    "FlatVectorFieldSchema",
+    "StorageType",
+    "IndexModel",
+    "FieldsModel",
+    "Schema",
+    "SchemaGenerator"
+]
